refactor(workspace): move ingest to breddit app, delete dead sqlite modules
Wave 2: Move ingest/ and snakify.ts to apps/breddit/src/lib/workspace/ingest/.
Remove ./ingest and ./ingest/reddit subpath exports from workspace package.json.
@sindresorhus/slugify stays in workspace (used by markdown materializer).

Wave 3: Delete extensions/materializer/sqlite/ (6 files) and
extensions/persistence/sqlite.ts. Remove ./extensions/persistence/sqlite
subpath export. drizzle-orm stays (re-exported from workspace root barrel).

613 tests pass, 0 fail.
