+++ apps/breddit/package.json
@@ -1,0 +1,24 @@
+{
+	"name": "@epicenter/breddit",
+	"description": "Reddit GDPR data export importer for Epicenter workspaces",
+	"private": true,
+	"version": "0.0.1",
+	"type": "module",
+	"exports": {
+		"./workspace": "./src/lib/workspace/ingest/index.ts"
+	},
+	"scripts": {
+		"typecheck": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@epicenter/workspace": "workspace:*",
+		"@sindresorhus/slugify": "catalog:",
+		"arktype": "catalog:",
+		"fflate": "catalog:"
+	},
+	"devDependencies": {
+		"bun-types": "catalog:",
+		"typescript": "catalog:"
+	},
+	"license": "MIT"
+}
