+++ apps/breddit/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"types": ["bun-types"]
+	},
+	"include": ["src"]
+}
